Don't fire empty searches
The representation linker uses a setTimeout to fire the typeahead
search, but the underlying model (this.text) might have changed
between when we test its length and when we fire that search.  Copy
the value upfront so we're working with a stable value.

Changed the linker to do the same thing ,even though right now there's
no timeout.  Later there probably will be.